@@ -1,9 +1,9 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2025 Surasia */
+use infinite_rs::TagStructure;
 use infinite_rs::tag::types::common_types::{
     AnyTag, FieldReal, FieldRealRGBColor, FieldRealVector2D, FieldReference,
 };
-use infinite_rs::TagStructure;
 
 #[derive(Debug, Default, TagStructure)]
 #[data(size(0xD8))]
