@@ -1,4 +1,6 @@
-use infinite_rs::{tag::types::common_types::FieldReference, TagStructure};
+/* SPDX-License-Identifier: GPL-3.0-or-later */
+/* Copyright © 2025 Surasia */
+use infinite_rs::{TagStructure, tag::types::common_types::FieldReference};
 
 #[derive(Default, Debug, TagStructure)]
 #[data(size(0x438))]
