@@ -1,6 +1,8 @@
+/* SPDX-License-Identifier: GPL-3.0-or-later */
+/* Copyright © 2025 Surasia */
 use infinite_rs::{
-    tag::types::common_types::{AnyTag, FieldBlock, FieldReference, FieldStringId},
     TagStructure,
+    tag::types::common_types::{AnyTag, FieldBlock, FieldReference, FieldStringId},
 };
 
 #[derive(Default, Debug, TagStructure)]
