@@ -1,11 +1,11 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2025 Surasia */
 use infinite_rs::{
+    TagStructure,
     tag::types::common_types::{
         FieldBlock, FieldRealBounds, FieldRealPoint3D, FieldRealVector3D, FieldReference,
         FieldShortInteger, FieldWordFlags,
     },
-    TagStructure,
 };
 
 use super::render_model::{MaterialBlock, MeshFlags};
