@@ -1,10 +1,10 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2025 Surasia */
 use infinite_rs::{
+    TagStructure,
     tag::types::common_types::{
         AnyTag, FieldBlock, FieldRealRGBColor, FieldReference, FieldStringId,
     },
-    TagStructure,
 };
 
 #[derive(Debug, Default, TagStructure)]
